+++ Mascoshop/Mascoshop/Mascoshop/src/main/java/com/Mascoshop/Entidades/ListaDeseos.java
@@ -17,7 +17,7 @@ public class ListaDeseos {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idListaDeseos;
+    private Integer idListaDeseos;
 
     @ManyToOne
     @JoinColumn(name = "idUsuario",nullable = false)
