+++ Mascoshop/Mascoshop/Mascoshop/src/main/java/com/Mascoshop/Entidades/ListaDeseosItem.java
@@ -1,6 +1,8 @@
 package com.Mascoshop.Entidades;
 
 import javax.persistence.*;
+
+import javax.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class ListaDeseosItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "idListaDeseos", nullable = false)
